+++ packages/http-client-csharp/generator/Microsoft.Generator.CSharp.ClientModel/src/Providers/RestClientProvider.cs
@@ -36,12 +36,14 @@ namespace Microsoft.Generator.CSharp.ClientModel.Providers
 
         private FieldProvider _pipelineMessageClassifier200;
         private FieldProvider _pipelineMessageClassifier201;
+        private FieldProvider _pipelineMessageClassifier202;
         private FieldProvider _pipelineMessageClassifier204;
         private FieldProvider _pipelineMessageClassifier2xxAnd4xx;
         private TypeProvider _classifier2xxAnd4xxDefinition;
 
         private PropertyProvider _classifier201Property;
         private PropertyProvider _classifier200Property;
+        private PropertyProvider _classifier202Property;
         private PropertyProvider _classifier204Property;
         private PropertyProvider _classifier2xxAnd4xxProperty;
 
@@ -51,11 +53,13 @@ namespace Microsoft.Generator.CSharp.ClientModel.Providers
             ClientProvider = clientProvider;
             _pipelineMessageClassifier200 = new FieldProvider(FieldModifiers.Private | FieldModifiers.Static, ClientModelPlugin.Instance.TypeFactory.StatusCodeClassifierApi.ResponseClassifierType, "_pipelineMessageClassifier200", this);
             _pipelineMessageClassifier201 = new FieldProvider(FieldModifiers.Private | FieldModifiers.Static, ClientModelPlugin.Instance.TypeFactory.StatusCodeClassifierApi.ResponseClassifierType, "_pipelineMessageClassifier201", this);
+            _pipelineMessageClassifier202 = new FieldProvider(FieldModifiers.Private | FieldModifiers.Static, ClientModelPlugin.Instance.TypeFactory.StatusCodeClassifierApi.ResponseClassifierType, "_pipelineMessageClassifier202", this);
             _pipelineMessageClassifier204 = new FieldProvider(FieldModifiers.Private | FieldModifiers.Static, ClientModelPlugin.Instance.TypeFactory.StatusCodeClassifierApi.ResponseClassifierType, "_pipelineMessageClassifier204", this);
             _classifier2xxAnd4xxDefinition = new Classifier2xxAnd4xxDefinition(this);
             _pipelineMessageClassifier2xxAnd4xx = new FieldProvider(FieldModifiers.Private | FieldModifiers.Static, _classifier2xxAnd4xxDefinition.Type, "_pipelineMessageClassifier2xxAnd4xx", this);
             _classifier200Property = GetResponseClassifierProperty(_pipelineMessageClassifier200, 200);
             _classifier201Property = GetResponseClassifierProperty(_pipelineMessageClassifier201, 201);
+            _classifier202Property = GetResponseClassifierProperty(_pipelineMessageClassifier202, 202);
             _classifier204Property = GetResponseClassifierProperty(_pipelineMessageClassifier204, 204);
             _classifier2xxAnd4xxProperty = new PropertyProvider(
                 $"Gets the PipelineMessageClassifier2xxAnd4xx",
@@ -76,6 +80,7 @@ namespace Microsoft.Generator.CSharp.ClientModel.Providers
             [
                 _classifier200Property,
                 _classifier201Property,
+                _classifier202Property,
                 _classifier204Property,
                 _classifier2xxAnd4xxProperty
             ];
@@ -99,6 +104,7 @@ namespace Microsoft.Generator.CSharp.ClientModel.Providers
             [
                 _pipelineMessageClassifier200,
                 _pipelineMessageClassifier201,
+                _pipelineMessageClassifier202,
                 _pipelineMessageClassifier204,
                 _pipelineMessageClassifier2xxAnd4xx
             ];
@@ -134,12 +140,8 @@ namespace Microsoft.Generator.CSharp.ClientModel.Providers
             var pipelineField = ClientProvider.PipelineProperty.ToApi<ClientPipelineApi>();
 
             var options = ScmKnownParameters.RequestOptions;
-            var parameters = GetMethodParameters(operation, true);
-            // All the parameters should be required for the CreateRequest method
-            foreach (var parameter in parameters)
-            {
-                parameter.DefaultValue = null;
-            }
+            var parameters = GetMethodParameters(operation, MethodType.CreateRequest);
+
             var signature = new MethodSignature(
                 $"Create{operation.Name.ToCleanName()}Request",
                 null,
@@ -164,7 +166,7 @@ namespace Microsoft.Generator.CSharp.ClientModel.Providers
                     message.ApplyResponseClassifier(classifier.ToApi<StatusCodeClassifierApi>()),
                     Declare("request", message.Request().ToApi<HttpRequestApi>(), out HttpRequestApi request),
                     request.SetMethod(operation.HttpMethod),
-                    Declare("uri", New.Instance<ClientUriBuilderDefinition>(), out ScopedApi<ClientUriBuilderDefinition> uri),
+                    Declare("uri", New.Instance(request.UriBuilderType), out ScopedApi uri),
                     uri.Reset(ClientProvider.EndpointField).Terminate(),
                     .. AppendPathParameters(uri, operation, paramMap),
                     .. AppendQueryParameters(uri, operation, paramMap),
@@ -179,7 +181,8 @@ namespace Microsoft.Generator.CSharp.ClientModel.Providers
 
         private IReadOnlyList<MethodBodyStatement> GetSetContent(HttpRequestApi request, IReadOnlyList<ParameterProvider> parameters)
         {
-            var contentParam = parameters.FirstOrDefault(p => ReferenceEquals(p, ScmKnownParameters.RequestContent));
+            var contentParam = parameters.FirstOrDefault(
+                p => ReferenceEquals(p, ScmKnownParameters.RequestContent) || ReferenceEquals(p, ScmKnownParameters.OptionalRequestContent));
             return contentParam is null ? [] : [request.Content().Assign(contentParam).Terminate()];
         }
 
@@ -196,6 +199,7 @@ namespace Microsoft.Generator.CSharp.ClientModel.Providers
                 {
                     200 => _classifier200Property,
                     201 => _classifier201Property,
+                    202 => _classifier202Property,
                     204 => _classifier204Property,
                     _ => throw new InvalidOperationException($"Unexpected status code {response.StatusCodes[0]}")
                 };
@@ -234,7 +238,7 @@ namespace Microsoft.Generator.CSharp.ClientModel.Providers
             return statements;
         }
 
-        private IEnumerable<MethodBodyStatement> AppendQueryParameters(ScopedApi<ClientUriBuilderDefinition> uri, InputOperation operation, Dictionary<string, ParameterProvider> paramMap)
+        private static List<MethodBodyStatement> AppendQueryParameters(ScopedApi uri, InputOperation operation, Dictionary<string, ParameterProvider> paramMap)
         {
             List<MethodBodyStatement> statements = new(operation.Parameters.Count);
 
@@ -245,14 +249,51 @@ namespace Microsoft.Generator.CSharp.ClientModel.Providers
 
                 string? format;
                 ValueExpression valueExpression;
-                GetParamInfo(paramMap, operation, inputParameter, out var type, out format, out valueExpression);
+                GetParamInfo(paramMap, operation, inputParameter, out var paramType, out format, out valueExpression);
                 var convertToStringExpression = TypeFormattersSnippets.ConvertToString(valueExpression, Literal(format));
-                ValueExpression toStringExpression = type?.Equals(typeof(string)) == true ? valueExpression : convertToStringExpression;
+                ValueExpression toStringExpression = paramType?.Equals(typeof(string)) == true ? valueExpression : convertToStringExpression;
                 MethodBodyStatement statement;
-                if (type?.IsCollection == true)
+
+                if (paramType?.IsCollection == true)
                 {
-                    statement = uri.AppendQueryDelimited(Literal(inputParameter.NameInRequest),
-                       valueExpression, format, true).Terminate();
+                    var delimiter = inputParameter.ArraySerializationDelimiter;
+
+                    if (inputParameter.Type is InputDictionaryType)
+                    {
+                        if (inputParameter.Explode)
+                        {
+                            statement = new ForeachStatement("param", valueExpression.AsDictionary(paramType),
+                                out KeyValuePairExpression item)
+                            {
+                                uri.AppendQuery(item.Key, item.Value, true).Terminate()
+                            };
+                        }
+                        else
+                        {
+                            statement = new[]
+                            {
+                                Declare("list", New.List<object>(), out var list),
+                                new ForeachStatement("param", valueExpression.AsDictionary(paramType), out KeyValuePairExpression item)
+                                {
+                                    list.Add(item.Key),
+                                    list.Add(item.Value)
+                                },
+                                uri.AppendQueryDelimited(Literal(inputParameter.NameInRequest), list, format, true)
+                                    .Terminate()
+                            };
+                        }
+                    }
+                    else if (!inputParameter.Explode)
+                    {
+                        statement = uri.AppendQueryDelimited(Literal(inputParameter.NameInRequest), valueExpression, format, true, delimiter: delimiter).Terminate();
+                    }
+                    else
+                    {
+                        statement = new ForeachStatement("param", valueExpression.As(paramType), out VariableExpression item)
+                        {
+                            uri.AppendQuery(Literal(inputParameter.NameInRequest), item, true).Terminate()
+                        };
+                    }
                 }
                 else
                 {
@@ -260,19 +301,55 @@ namespace Microsoft.Generator.CSharp.ClientModel.Providers
                         .Terminate();
                 }
 
-                statement = inputParameter.IsRequired
-                    ? statement
-                    : new IfStatement(valueExpression.NotEqual(Null))
-                    {
-                        statement
-                    };
+                if (!inputParameter.IsRequired || paramType?.IsNullable == true ||
+                    (paramType is { IsValueType: false, IsFrameworkType: true } && paramType.FrameworkType != typeof(string)))
+                {
+                    statement = BuildQueryParameterNullCheck(paramType, valueExpression, statement);
+                }
+
                 statements.Add(statement);
             }
 
             return statements;
         }
 
-        private IEnumerable<MethodBodyStatement> AppendPathParameters(ScopedApi<ClientUriBuilderDefinition> uri, InputOperation operation, Dictionary<string, ParameterProvider> paramMap)
+        private static IfStatement BuildQueryParameterNullCheck(
+            CSharpType? parameterType,
+            ValueExpression valueExpression,
+            MethodBodyStatement originalStatement)
+        {
+            if (parameterType?.IsCollection == true)
+            {
+                DeclarationExpression? changeTrackingCollectionDeclaration;
+                VariableExpression? changeTrackingReference;
+                if (parameterType.IsDictionary)
+                {
+                    changeTrackingCollectionDeclaration = Declare(
+                        "changeTrackingDictionary",
+                        ClientModelPlugin.Instance.TypeFactory.DictionaryInitializationType.MakeGenericType(parameterType.Arguments),
+                        out changeTrackingReference);
+                }
+                else
+                {
+                    changeTrackingCollectionDeclaration = Declare(
+                        "changeTrackingList",
+                        ClientModelPlugin.Instance.TypeFactory.ListInitializationType.MakeGenericType(parameterType
+                            .Arguments),
+                        out changeTrackingReference);
+                }
+
+                return new IfStatement(valueExpression.NotEqual(Null)
+                    .And(Not(valueExpression.Is(changeTrackingCollectionDeclaration)
+                    .And(changeTrackingReference.Property("IsUndefined")))))
+                {
+                    originalStatement
+                };
+            }
+
+            return new IfStatement(valueExpression.NotEqual(Null)) { originalStatement };
+        }
+
+        private IEnumerable<MethodBodyStatement> AppendPathParameters(ScopedApi uri, InputOperation operation, Dictionary<string, ParameterProvider> paramMap)
         {
             Dictionary<string, InputParameter> inputParamHash = new(operation.Parameters.ToDictionary(p => p.Name));
             List<MethodBodyStatement> statements = new(operation.Parameters.Count);
@@ -286,7 +363,7 @@ namespace Microsoft.Generator.CSharp.ClientModel.Providers
         private void AddUriSegments(
             string segments,
             int offset,
-            ScopedApi<ClientUriBuilderDefinition> uri,
+            ScopedApi uri,
             List<MethodBodyStatement> statements,
             Dictionary<string, InputParameter> inputParamHash,
             Dictionary<string, ParameterProvider> paramMap,
@@ -313,13 +390,14 @@ namespace Microsoft.Generator.CSharp.ClientModel.Providers
                 CSharpType? type;
                 string? format;
                 ValueExpression valueExpression;
+                InputParameter? inputParam = null;
                 if (isClientParameter)
                 {
                     GetParamInfo(paramMap[paramName], out type, out format, out valueExpression);
                 }
                 else
                 {
-                    var inputParam = inputParamHash[paramName];
+                    inputParam = inputParamHash[paramName];
                     if (inputParam.Location == RequestLocation.Path || inputParam.Location == RequestLocation.Uri)
                     {
                         GetParamInfo(paramMap, operation, inputParam, out type, out format, out valueExpression);
@@ -330,8 +408,19 @@ namespace Microsoft.Generator.CSharp.ClientModel.Providers
                     }
                 }
                 ValueExpression[] toStringParams = format is null ? [] : [Literal(format)];
-                valueExpression = type?.Equals(typeof(string)) == true ? valueExpression : valueExpression.Invoke(nameof(ToString), toStringParams);
-                statements.Add(uri.AppendPath(valueExpression, true).Terminate());
+                bool escape = !inputParam?.SkipUrlEncoding ?? true;
+                if (type?.OutputType.IsCollection == true)
+                {
+                    statements.Add(uri.AppendPathDelimited(
+                        valueExpression, format, escape, inputParam?.ArraySerializationDelimiter).Terminate());
+                }
+                else
+                {
+                    valueExpression = type?.Equals(typeof(string)) == true
+                        ? valueExpression
+                        : valueExpression.Invoke(nameof(ToString), toStringParams);
+                    statements.Add(uri.AppendPath(valueExpression, escape).Terminate());
+                }
 
                 pathSpan = pathSpan.Slice(paramEndIndex + 1);
             }
@@ -425,7 +514,7 @@ namespace Microsoft.Generator.CSharp.ClientModel.Providers
             return MethodCache[operation];
         }
 
-        internal static List<ParameterProvider> GetMethodParameters(InputOperation operation, bool isProtocol = false)
+        internal static List<ParameterProvider> GetMethodParameters(InputOperation operation, MethodType methodType)
         {
             SortedList<int, ParameterProvider> sortedParams = [];
             int path = 0;
@@ -445,11 +534,20 @@ namespace Microsoft.Generator.CSharp.ClientModel.Providers
 
                 ParameterProvider? parameter = ClientModelPlugin.Instance.TypeFactory.CreateParameter(inputParam).ToPublicInputParameter();
 
-                if (isProtocol)
+                if (methodType is MethodType.Protocol or MethodType.CreateRequest)
                 {
                     if (inputParam.Location == RequestLocation.Body)
                     {
-                        parameter = ScmKnownParameters.RequestContent;
+                        if (methodType == MethodType.CreateRequest)
+                        {
+                            parameter = ScmKnownParameters.RequestContent;
+                        }
+                        else
+                        {
+                            parameter = parameter.DefaultValue == null
+                                ? ScmKnownParameters.RequestContent
+                                : ScmKnownParameters.OptionalRequestContent;
+                        }
                     }
                     else
                     {
@@ -510,6 +608,15 @@ namespace Microsoft.Generator.CSharp.ClientModel.Providers
                 sortedParams.Add(bodyRequired++, ScmKnownParameters.ContentType);
             }
 
+            // All the parameters should be required for the CreateRequest method
+            if (methodType == MethodType.CreateRequest)
+            {
+                foreach (var parameter in sortedParams.Values)
+                {
+                    parameter.DefaultValue = null;
+                }
+            }
+
             return [.. sortedParams.Values];
         }
 
@@ -521,6 +628,13 @@ namespace Microsoft.Generator.CSharp.ClientModel.Providers
             }
 
             throw new InvalidOperationException($"inputParam `{inputParam.Name}` is `Spread` but not a model type");
+        }
+
+        internal enum MethodType
+        {
+            CreateRequest,
+            Protocol,
+            Convenience
         }
     }
 }