+++ packages/http-client-csharp/generator/Microsoft.Generator.CSharp.ClientModel/test/Providers/Abstractions/HttpMessageApiTests.cs
@@ -6,7 +6,6 @@ using Microsoft.Generator.CSharp.Primitives;
 using Microsoft.Generator.CSharp.Statements;
 using Microsoft.Generator.CSharp.Tests.Common;
 using NUnit.Framework;
-using static Microsoft.Generator.CSharp.Snippets.Snippet;
 
 namespace Microsoft.Generator.CSharp.ClientModel.Tests.Providers.Abstractions
 {
@@ -59,9 +58,6 @@ namespace Microsoft.Generator.CSharp.ClientModel.Tests.Providers.Abstractions
 
             public override ValueExpression BufferResponse()
                 => Original.Invoke("GetFakeBufferResponse");
-
-            public override MethodBodyStatement[] ExtractResponse()
-                => [Return(Original.Invoke("GetFakeExtractResponse"))];
 
             public override HttpMessageApi FromExpression(ValueExpression original)
                 => new TestHttpMessageApi(original);