+++ packages/http-client-csharp/generator/Microsoft.Generator.CSharp.ClientModel/test/Providers/ClientProviders/ClientProviderTests.cs
@@ -361,17 +361,18 @@ namespace Microsoft.Generator.CSharp.ClientModel.Tests.Providers.ClientProviders
             Assert.AreEqual(2, methods.Where(m => m.Signature.Parameters.Any(p => p.Name == "queryParam" && p.Type.IsFrameworkType && p.Type.FrameworkType == typeof(string))).Count());
         }
 
-        [Test]
-        public void ValidateQueryParamWriterDiff()
+        [TestCase(true)]
+        [TestCase(false)]
+        public void ValidateQueryParamWriterDiff(bool isAsync)
         {
             MockHelpers.LoadMockPlugin(
-                createClientCore: (client) => new ValidateQueryParamDiffClientProvider(client));
+                createClientCore: (client) => new ValidateQueryParamDiffClientProvider(client, isAsync));
 
             var clientProvider = ClientModelPlugin.Instance.TypeFactory.CreateClient(GetEnumQueryParamClient());
 
             TypeProviderWriter writer = new(clientProvider);
             var codeFile = writer.Write();
-            Assert.AreEqual(Helpers.GetExpectedFromFile(), codeFile.Content);
+            Assert.AreEqual(Helpers.GetExpectedFromFile(isAsync.ToString()), codeFile.Content);
         }
 
         [Test]
@@ -448,23 +449,25 @@ namespace Microsoft.Generator.CSharp.ClientModel.Tests.Providers.ClientProviders
             Assert.IsNotNull(requestOptionsParameterInAsyncMethod);
             Assert.AreEqual(shouldBeOptional, requestOptionsParameterInAsyncMethod!.Type.IsNullable);
 
+            // request options should always be last parameter
+            Assert.AreEqual("RequestOptions", syncMethod.Signature.Parameters[^1].Type.Name);
+            Assert.AreEqual("RequestOptions", asyncMethod.Signature.Parameters[^1].Type.Name);
+
             if (shouldBeOptional)
             {
                 Assert.IsNotNull(requestOptionsParameterInSyncMethod.DefaultValue);
                 Assert.IsNotNull(requestOptionsParameterInAsyncMethod.DefaultValue);
-                // request options should be last optional parameter
-                Assert.AreEqual("RequestOptions", syncMethod.Signature.Parameters[^1].Type.Name);
-                Assert.AreEqual("RequestOptions", asyncMethod.Signature.Parameters[^1].Type.Name);
             }
-            else
+
+            if (shouldBeOptional && hasOptionalParameter)
             {
-                Assert.IsNull(requestOptionsParameterInSyncMethod.DefaultValue);
-                Assert.IsNull(requestOptionsParameterInAsyncMethod.DefaultValue);
-                // optional parameters should come after the required request options
-                if (hasOptionalParameter)
+                var optionalParameter = syncMethod.Signature.Parameters[^2];
+                // The optional parameter should be required in protocol method
+                Assert.IsNull(optionalParameter.DefaultValue);
+                // It should also be nullable for value types
+                if (optionalParameter.Type.IsValueType)
                 {
-                    Assert.AreEqual("RequestOptions", syncMethod.Signature.Parameters[^2].Type.Name);
-                    Assert.AreEqual("RequestOptions", asyncMethod.Signature.Parameters[^2].Type.Name);
+                    Assert.IsTrue(optionalParameter.Type.IsNullable);
                 }
             }
         }
@@ -572,14 +575,19 @@ namespace Microsoft.Generator.CSharp.ClientModel.Tests.Providers.ClientProviders
 
         private class ValidateQueryParamDiffClientProvider : ClientProvider
         {
-            public ValidateQueryParamDiffClientProvider(InputClient client)
+            private readonly bool _isAsync;
+
+            public ValidateQueryParamDiffClientProvider(InputClient client, bool isAsync = false)
                 : base(client)
             {
+                _isAsync = isAsync;
             }
 
             protected override MethodProvider[] BuildMethods()
             {
-                var method = base.BuildMethods().Where(m => m.Signature.Parameters.Any(p => p.Name == "queryParam" && p.Type.Name == "InputEnum" && !m.Signature.Name.EndsWith("Async"))).First();
+                var method = base.BuildMethods().First(m => m.Signature.Parameters.Any(p =>
+                    p is { Name: "queryParam", Type.Name: "InputEnum" } &&
+                    ((_isAsync && m.Signature.Name.EndsWith("Async")) || (!_isAsync && !m.Signature.Name.EndsWith("Async")))));
                 method.Update(xmlDocProvider: new XmlDocProvider()); // null out the docs
                 return [method];
             }
@@ -736,8 +744,8 @@ namespace Microsoft.Generator.CSharp.ClientModel.Tests.Providers.ClientProviders
                                 isRequired: true),
                         ]), false, false);
 
-                // Protocol & convenience methods will have the same parameters, so RequestOptions should be required.
-                // One of the parameter is optional, so RequestOptions should come before it
+                // Protocol & convenience methods will have the same parameters.
+                // One of the parameter is optional, so it should be made required in the protocol method.
                 yield return new TestCaseData(
                     InputFactory.Operation(
                         "TestOperation",
@@ -746,6 +754,25 @@ namespace Microsoft.Generator.CSharp.ClientModel.Tests.Providers.ClientProviders
                             InputFactory.Parameter(
                                 "p1",
                                 InputPrimitiveType.String,
+                                location: RequestLocation.None,
+                                isRequired: false),
+                            InputFactory.Parameter(
+                                "p2",
+                                InputPrimitiveType.Int64,
+                                location: RequestLocation.None,
+                                isRequired: true),
+                        ]), false, true);
+
+                // Protocol & convenience methods will have the same parameters.
+                // One of the parameter is optional value type, so it should be made nullable required in the protocol method, and RequestOptions can be optional.
+                yield return new TestCaseData(
+                    InputFactory.Operation(
+                        "TestOperation",
+                        parameters:
+                        [
+                            InputFactory.Parameter(
+                                "p1",
+                                InputPrimitiveType.Int32,
                                 location: RequestLocation.None,
                                 isRequired: false),
                             InputFactory.Parameter(