+++ packages/http-client-java/emitter/src/emitter.ts
@@ -12,6 +12,7 @@ import { dump } from "js-yaml";
 import { dirname } from "path";
 import { fileURLToPath } from "url";
 import { CodeModelBuilder } from "./code-model-builder.js";
+import { logError } from "./utils.js";
 
 export interface EmitterOptions {
   namespace?: string;
@@ -41,6 +42,8 @@ export interface EmitterOptions {
 
   "group-etag-headers"?: boolean;
 
+  "enable-subclient"?: boolean;
+
   "advanced-versioning"?: boolean;
   "api-version"?: string;
   "service-version-exclude-preview"?: boolean;
@@ -63,7 +66,7 @@ const EmitterOptionsSchema: JSONSchemaType<EmitterOptions> = {
     "output-dir": { type: "string", nullable: true },
     "package-dir": { type: "string", nullable: true },
 
-    flavor: { type: "string", nullable: true, default: "Azure" },
+    flavor: { type: "string", nullable: true },
 
     // service
     "service-name": { type: "string", nullable: true },
@@ -78,7 +81,7 @@ const EmitterOptionsSchema: JSONSchemaType<EmitterOptions> = {
 
     "enable-sync-stack": { type: "boolean", nullable: true, default: true },
     "stream-style-serialization": { type: "boolean", nullable: true, default: true },
-    "use-object-for-unknown": { type: "boolean", nullable: true, default: true },
+    "use-object-for-unknown": { type: "boolean", nullable: true, default: false },
 
     // customization
     "partial-update": { type: "boolean", nullable: true, default: false },
@@ -89,6 +92,8 @@ const EmitterOptionsSchema: JSONSchemaType<EmitterOptions> = {
     polling: { type: "object", additionalProperties: true, nullable: true },
 
     "group-etag-headers": { type: "boolean", nullable: true },
+
+    "enable-subclient": { type: "boolean", nullable: true, default: false },
 
     "advanced-versioning": { type: "boolean", nullable: true, default: false },
     "api-version": { type: "string", nullable: true },
@@ -113,10 +118,7 @@ export async function $onEmit(context: EmitContext<EmitterOptions>) {
   if (!options["flavor"]) {
     if (options["package-dir"]?.toLocaleLowerCase().startsWith("azure")) {
       // Azure package
-      options["flavor"] = "Azure";
-    } else {
-      // default
-      options["flavor"] = "Azure";
+      options["flavor"] = "azure";
     }
   }
   const builder = new CodeModelBuilder(program, context);
@@ -135,7 +137,8 @@ export async function $onEmit(context: EmitContext<EmitterOptions>) {
 
     await promises.mkdir(outputPath, { recursive: true }).catch((err) => {
       if (err.code !== "EISDIR" && err.code !== "EEXIST") {
-        throw err;
+        logError(program, `Failed to create output directory: ${outputPath}`);
+        return;
       }
     });
 
@@ -232,9 +235,9 @@ export async function $onEmit(context: EmitContext<EmitterOptions>) {
           message: msg,
           target: NoTarget,
         });
-        throw new Error(msg);
+        logError(program, msg);
       } else {
-        throw error;
+        logError(program, error.message);
       }
     }