+++ packages/http-client-python/CHANGELOG.md
@@ -1,5 +1,27 @@
 # Change Log - @typespec/http-client-python
 
+## 0.3.11
+
+### Other Changes
+
+- Pad special property name in model to avoid conflict
+
+### Bug Fixes
+
+- Fix crash if no valid client define in typespec file
+
+## 0.3.10
+
+### Bug Fixes
+
+- Bump pyright and mypy dependencies
+
+## 0.3.9
+
+### Bug Fixes
+
+- Fix quote for string type
+
 ## 0.3.8
 
 ### Bug Fixes